feat: order 기능 추가 및 OpenFeign, Kafka 설정

+++ backend/order/src/main/java/com/zzimcong/order/application/dto/OrderItemRequest.java
@@ -3,10 +3,12 @@ package com.zzimcong.order.application.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class OrderItemRequest {
     public Long productId;
-    public Integer price;
+    public BigDecimal price;
     public Integer quantity;
 }
